#1656 split up scram_over_mtls module
